⏫ upgrade dependencies
This updates the crate's dependencies, most notably the `tokio` version
used in tests. Note that `prometheus` is NOT upgraded, because its types
are included in the public API, and thus entails opening a separate can
of worms.

These updates are purely intended to be polish while we're in the
neighborhood, i.e. `^` semver carets, alphabetizing, etc.