😬 revert pinning changes
This reverts the changes in c27dd7d and 3a1fd36. Unfortunately, #3 was
founded on some slight misunderstandings of `Pin` semantics — namely,
that our underlying future _will_ be pinned when an instrumented future
is itself pinned.

`pin_project` will generate code that prevents UB, and when an
instrumented future is `.await`ed, the executor will pin it for us.
Because pinning is structural for the inner future `F`, it too will be
pinned and our poll calls will function properly.

Apologies for the false alarms, but I am happy to know that this is (and
was) functioning like it should.