🛑 guards: deferred increments can be cancelled
 ### ⛅ background

the `deferred_inc` and `deferred_add` interfaces in this library can be
used to postpone the act of updating a particular Prometheus metric.

`complete_add` provides a way to do this /explicitly/, but the `Drop`
implementation will do this /implicitly/.

however, once one of these RAII guards are created, there is not a way
to opt-out of this. `std::mem::forget` may be tempting here, but in the
case of `DeferredAddWithLabels`, we shouldn't leak our `L` value.

 ### ➕ changes

this commit adds a `cancel` method to `DeferredAddWithLabels` and
`DeferredAdd`. if invoked, the guard will be consumed and no change will
be made to the metric in question.

note that to facilitate this, the inner value is now an `Option<T>`. the
`Drop` code will only run when this is `Some(value)`. the fact that this
is an optional is not exposed in the public interface, `cancel` is
the only way to "take" that value.